[#2] correções resolvidas
