+++ lib/app/modules/home/data/service/news_services.dart
@@ -5,11 +5,18 @@ import 'package:intoxianime/app/modules/home/data/model/news_model.dart';
 class NewsService {
   final Dio dio;
 
-  NewsService(this.dio);
+  const NewsService(this.dio);
+
+  static const String _baseUrl = 'https://www.intoxianime.com/';
 
   Future<List<NewsModel>> getAllNews() async {
     final response = await dio.get(
-      'https://www.intoxianime.com/?rest_route=/wp/v2/posts&page=1&per_page=10',
+      _baseUrl,
+      queryParameters: {
+        'rest_route': '/wp/v2/posts',
+        'page': 1,
+        'per_page': 10,
+      },
     );
     return NewsAdapter.fromList(response.data);
   }
